Initial attempt at Ti.UI.Clipboard implementation for iOS, following the Titanium Desktop version's API.
setText(), hasText(), and getText() work and are successfully demo'd in KitchenSink (separate commit).

Data untested, but in theory should allow sending/receiving images (from image, data, or file-backed blobs) and raw-data blobs, though there may not be proper mapping of MIME types to UTIs for the iOS pasteboard.
Special handling of 'image' and 'url' types to explicitly mark as image or URL, as per the Desktop version.


Former-commit-id: 46c00d6fe8ddc046b71cf0252450b8c959e9155f