[#2654 state:fixed-in-qa][#1058 state:fixed-in-qa] update HttpClient to support app wide HttpClient instance
Modified TiHTTPClient to be a static member so that the client values such as cookies will live over multiple requests.  Also modifed the credentials to now pay attention to the url scope they are assigned to and not used globally.  Made an additional change to set the source property for HttpClient callbacks.


Former-commit-id: 907b19d1aff9dcdd2468df1af9a5ae64ba335918