[#1386 state:fixed-in-qa] update geolocation module to switch providers
Updated the geolocation module to check for updated location provider with each new location update.  While not a perfect solution for handling the switching of the location providers, this will allow the provider to switch based on availability without forcing the user to register again.  Further discussion is needed regarding a more complete solution for handling location switching and response time.  Also modified geolocation to completely decouple the TiLocationHelper / TiCompassHelper from the proxy.  Updated KS geolocation example to set gps as the preferredProvider and correct a bug in the pause / resume handling of location event listeners.


Former-commit-id: f9036efc1f4a524603bffb8bb5ad759667a339d0