[#3101 state:fixed-in-qa] [#3100 state:in-progress] update orientation mechanism to use tighter shift band
updated orientation mechanism to use tighter shift badn of +- 10 degrees off center of new orientation to cause orientation shift.  Also updated logic to disable setting orientation to an orientation that is not listed in the orientationModes.  #3100 is still in progress because it requires more discussion with iOS team during planning week to decide on a unified long term approach to manually setting orientation change.


Former-commit-id: d55abb0d6cc6be9d8803239e7ebd028fdc1a3702