New property "autoRedirect" for Ti.Network.HTTPClient
This commit introduces the new property autoRedirect to Ti.Network.HTTPClient for both android and iphone. By default it is true, aligning with the current behaviour, so that Titanium handles the HTTP redirect.

Since however there are many problems with handling of redirections (for exampleTIMOB-3360) I prefer handling the redirect behaviour in the Javascript code instead. Therefore if autoRedirect is set to false then the HTTPClient simple returns the 30x response along with all the headers.


Former-commit-id: ef403ac1a673986ba53a2dd8105b76e6f979afe3