[TIMOB-5327] For submodules, we need to rememberProxy: - what follows is incredibly detailed information on the bug:
This bug was difficult to track down, and intermittent, because it relies entirely
on where the submodules were situated in JSCore heap blocks. It turns out that JSCore's
sweep of unmarked (read: 'unprotected') cells destroys them and then calls a placement
new() on them to force the constructor to dump the null/undefined structure into the
memory block. But then, for whatever reason, this would not always cause an
[undefined] access when querying the submodule a second time, after GC. It appears
that there is a second component to detecting bugs like this one, which involves
bad bridge interaction (obj-c counts not matching protect counts). It could be
the case that we're GCing different heaps at different times, which would
be some bad news (too much garbage resident in memory).

Short version: If you pull an [undefined] value where a proxy should be, it's
because it was unprotected at some point.


Former-commit-id: 2c451bab0a161ab85fec171e323cd6a7da03c99c