Correcting scroll events for ScrollableView.
Instead of using the `onPageSelected` handler of our
`ViewPager.SimpleOnPageChangeListener` to fire `dragEnd` events, use the
`onPageScrollStateChanged`.  This will properly fire `dragEnd` events
even when we haven't actually completed a scroll to another view.

Instead of using the `position` parameter passed to the `onPageScrolled`
handler of our `ViewPager.SimpleOnPageChangeListener` alone to determine the
current index, use it in combination with the `positionOffset`;
rounding.  This way, we get a value of 2 for the index when we are more
than halfway scrolled onto page 2.


Former-commit-id: 9a758f9554f2cc8bc0b5963114083fc1550033fe