Fixing paging indicators & edge case of scrollEnd.
The paging indicators broke because I mistakenly made a property
`mCurIndex` of the page change listener for the `ViewPager`.  Thus, the
indicators didn't know what page to go to when clicked.  I was also
checking to see if they needed to be shown at the wrong time; this has
been corrected by calling `showPager()` whenever we fire a `scrollEnd`.

This commit also introduces a hacky fix for an edge case of `scrollEnd`.
If the user drags the view and manually settles it on another view, the
`onPageScrollStateChanged` handler is never called.  So, we use the
`onPageSelected` handler in this case; but only if a `dragEnd` event
wasn't fired beforehand.


Former-commit-id: c8deb8f754fbcdbd5cc19caee6edeaa29e988a53