Add "screenSize" to the list of "configChanges" that our Activities handle.  This is necessary because beginning in api level 13 (Honeycomb MR2, Android 3.2), screenSize is not handled automatically by Android if the targetSdkVersion is set to 13 or higher.  The consequence of this is that if you change the orientation of the device (e.g., portrait->landscape), the activity will restart (unless you indicate in AndroidManifest.xml that your activity handles screenSize).  Since we do not support system restarts of activities, in Titanium this leads to the entire app shutting down and restarting.
By adding screenSize to the list of configChanges in our template AndroidManifest.xml, we can prevent the application re-start.

However, since the addition of screenSize will cause aapt to fail when building against versions of Android that are older than api level 13, we now also automatically build against the latest version found in the user's Android SDK folder.  And if we do not find a version that is at least 13, we remove screenSize from the generated AndroidManifest.xml so as to allow aapt to finish successfully.

[TIMOB-10863]


Former-commit-id: 7cbfe22807aeb322a2bed94b24969e3f44ddd49b