[TIMOB-768] Suppress pinch events until the pointers have moved a few pixels.
This allows twofingertap events to fire on Hieu's Droid X, and presumably other older
devices. This means that the first pinch event in a pinch gesture is delayed until the
total distance moved passes an arbitrary threshold (set at 6 pixels). After that, pinch
events proceed as usual until the end of the gesture.

If that delay is unacceptable, we could short-circuit this logic if no twofingertap
listener is defined.


Former-commit-id: 93602ccae112be3073acfcaf52cd9de958aacaaf