added contentSize property to scroll event
The 'scroll' event for [Titanium.UI.ScrollView](http://docs.appcelerator.com/titanium/latest/#!/api/Titanium.UI.ScrollView) does not send back the contentSize for the scroll view. Having the content size is useful when adding items to the bottom of the view based on user scrolling (nearing the end). Tableview does have this, in TiUITableView.m in eventObjectForScrollView.

Former-commit-id: d198a4fc506962b59fbbe3b9ffed6bcd5bfc15c8