[TC-5432] add support for ellipsize modes
This changes the type of `ellipsize` of TiUILabel to TruncateAt on the analogy of the native TextView. The ellipsize modes are exposed to
Ti.UI.TEXT_ELLIPSIZE_TRUNCATE_START
Ti.UI.TEXT_ELLIPSIZE_TRUNCATE_MIDDLE
Ti.UI.TEXT_ELLIPSIZE_TRUNCATE_END
Ti.UI.TEXT_ELLIPSIZE_TRUNCATE_MARQUEE

So as not to break anything I've added a check if `ellipsize` is a boolean falling back to the previous behavior.


Former-commit-id: a32d8e27be2e339530a5cd5ad33aa9a93fced535