Add _app_props to resolve warning in Xcode source
The `_app_props.json` is usually generated during our build process. Since the Xcode source project does not use that, a warning is thrown in the Xcode console that might confuse new developers who are trying to build the SDK-project from Xcode. 

In fact, the file looks like this:
```json
{"ti.ui.defaultunit":"dp","appc-app-id":"000000000000000000000000","ti.deploytype":"development"}
```
and only the `ti.ui.defaultunit` key is references in the project. Having this file in, gives us an even better reflection on how the project looks built.

Former-commit-id: 6eee4a921754b8a7e5110a1e259440611ec4218f