[TIMOB-4865] iOS: imageAsCropped rotates image
The issue with the camera returning images with the wrong orientation is a very old issue that I found on stack and apple forums. All of which had this as the only proposed fix.
[Apple Forum](https://discussions.apple.com/message/7271151#7271151)
[Stack 1](http://stackoverflow.com/questions/10600613/ios-image-orientation-has-strange-behavior)
[Stack 2](http://stackoverflow.com/questions/5427656/ios-uiimagepickercontroller-result-image-orientation-after-upload)

This change will not effect the behavior of any of our UI elements that can contain an image.


Former-commit-id: fd472543e01b5beefcce82357076d7f0adc3af2c