[TIMOB-18565] Adopt Android Library handling from Gradle
This will change the handling of Android Libraries to be more like
Gradle does it. Instead of pre-merging resources, copying .jar and .so
libraries from every Android Archive, and package that into the final
module build, we only extract the Android Archive and pass those files
to the appropriate module build steps. The Android Archive will then be
packaged as is into the module and the app builder can process the
archive again.


Former-commit-id: af5a92dc09e3e0b63ff8b8d3562c6402c2192830